Test with bigint fixed

+++ src/test/resources/test-schema.sql
@@ -2,7 +2,7 @@ DROP TABLE IF EXISTS wishlist;
 DROP TABLE IF EXISTS users;
 
 CREATE TABLE users (
-                       user_id INT PRIMARY KEY,
+                       user_id bigint PRIMARY KEY,
                        email VARCHAR(255) UNIQUE NOT NULL,
                        name VARCHAR(255) NOT NULL,
                        password VARCHAR(255) NOT NULL
@@ -13,5 +13,5 @@ CREATE TABLE wishlist (
                           name VARCHAR(255),
                           description VARCHAR(255),
                           image_url VARCHAR(255),
-                          user_id INT
+                          user_id bigint
 );
